refactor: 입출력 관련 로직을 AndroidView로 분리
